Issue 14: Organized Tests A.2 - Move unit tests to separate files (Part 2)
Moved additional unit tests from integration test files:

✅ test__children_selector.py - ChildrenSelector unit tests
   - Extracted TestChildrenSelector class from integration file
   - Kept TestChildrenSelectorIntegration in integration file

Changes:
- Created 1 new unit test file
- Separated unit tests from integration tests in children_selector
- Updated integration file to focus on adapter system integration
- All 528 tests still pass

Total unit test files created so far: 6
